refactor: ranking 관련 utils 파일 분리 #11
