+++ src/components/ranking/PlanDetailItem.tsx
@@ -3,7 +3,8 @@ import PlanTooltip from '@/components/ranking/PlanTooltip';
 import { Button } from '@/components/ui/button';
 import useRankDetailQuery from '@/hooks/queries/ranking/useRankDetailQuery';
 import { usePlanTooltip } from '@/hooks/ranking/usePlanTooltip';
-import { calculateDataComment } from '@/utils/planHelpers';
+import { calculateDataComment } from '@/utils/ranking/formatPlanData';
+
 interface PlanDetailItemProps {
   planId: number;
   onBack: () => void;
