+++ src/utils/ranking/formatPlanData.ts
@@ -1,10 +1,6 @@
-import {
-  getBasicBenefitDescription,
-  BENEFIT_DESCRIPTIONS,
-  DescriptionLine,
-} from '@/constants/benefitDescriptions';
-import { RankingPlanDetailResponse } from '@/types/ranking';
-import { getDataComment } from '@/utils/dataUsage';
+import type { RankingPlanDetailResponse } from '@/types/ranking';
+import { getDataComment } from '@/utils/ranking/calculateDataUsage';
+import { getBenefitDescription } from '@/utils/ranking/getBenefitDescription';
 
 // 데이터 코멘트 계산
 export const calculateDataComment = (planData: RankingPlanDetailResponse) => {
@@ -12,14 +8,6 @@ export const calculateDataComment = (planData: RankingPlanDetailResponse) => {
     dataType: planData.dataType,
     dataAmountGb: planData.dataAmountGb,
   });
-};
-
-// 🆕 혜택 설명 가져오기 (DescriptionLine[] 반환)
-export const getBenefitDescription = (key: string, value: string): DescriptionLine[] => {
-  if (key === '기본혜택') {
-    return getBasicBenefitDescription(value);
-  }
-  return BENEFIT_DESCRIPTIONS[key] || [];
 };
 
 // 안전한 혜택 값 가져오기
@@ -30,7 +18,7 @@ export const getBenefitValue = (allBenefits: any, key: string): string => {
   return benefits[key] || '';
 };
 
-// 🆕 설명이 있는지 확인 (DescriptionLine[] 길이 체크)
+// 설명이 있는지 확인
 export const hasDescription = (key: string, value: string): boolean => {
   const description = getBenefitDescription(key, value);
   return description.length > 0;
